src/Makefile: don't hard-code system lib dir
Currently, we add -L/usr/lib64 to LINK_FLAGS, and then add that to all
linker invocations.

However, there's no guarantee that /usr/lib64 contains libraries
suitable for this build, so remove it. It's up to the build environment
to point to the relevant lib paths through the standard LDFLAGS.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>