tree: rework nvme_scan_topology()
Rework nvme_scan_topology() to scan for controllers, and not
for subsystems. As the hostnqn is part of the controller structure
we cannot derive the host from the subsystem, and hence may end
up using the wrong host.

This patch updates to library version to 1.1, as it adds the two
new symbols 'nvme_configure_ctrl', 'nvme_deconfigure_ctrl', and
'nvme_scan_ctrls'.

Signed-off-by: Hannes Reinecke <hare@suse.de>