nvme: rework controller allocation
When connecting a fabrics controller we start off with a controller
entry, but then need to match it against the sysfs structure.
Creating an unconnected controller has the issue that another controller
entry might already be present, and we would need to merge them.
Which is not only awkward, but also error prone.

So rework the allocation scheme by always use nvme_lookup_ctrl(),
as even when connecting the controller we know the subsystem
(at least by name), and move the controller to the correct subsystem
after connection.

With that we can make nvme_create_ctrl() an internal symbol, and
increase the version to 1.0.2 to signal that change.

Signed-off-by: Hannes Reinecke <hare@suse.de>