Add Python tests to meson framework.
Python has this strict directory layout requirement. A Python
package is a directory that must contain a __init__.py file in
addition to containing modules.

To be able to run Python tests and load modules directly from the
build directory, the Python package (libnvme) and module (nvme.py)
must be organized in the build directory the same way they would
be in the install directory. And the PYTHONPATH must be set to
point to the build directory. In other words, like this:

.build              <- PYTHONPATH points to here.
 \_ libnvme         <- Package
     \_ nvme.py     <- Module
     \_ __init__.py
     \_ _nvme.cpython-38-x86_64-linux-gnu.so

This patch changes the directory layout so that meson will
organize the Python package and module as shown above. The patch
also defines Python tests and sets the PYTHONPATH to allow
importing directly from the build directory.

Signed-off-by: Martin Belanger <martin.belanger@dell.com>