nvme-cli: ctrl-loss-tmo should accept -1 as value
When passing the parameter --ctrl-loss-tmo=-1
nvme-cli silently drops it and replaces its value with the
default one (600) despite the fact that -1 is valid
and accepted by the kernel (it means "off" and
that it will try to reconnect forever).

This is due to the fact that -1 is treated like "variable unset"
by nvme-cli and the add_int_argument() function ignores
all negatives arguments.

This patch fixes the bug by allowing arguments to have
negative values and removing the if condition that assumes that
ctrl_loss_tmo = -1 means that the variable has not been set.

Signed-off-by: Maurizio Lombardi <mlombard@redhat.com>