Fix ccan mem-leaks
meson provide the option: -Db_sanitize=address.

Ex: meson .build -Db_sanitize=address && ninja -C .build

This is used to configure the build environment with the
gcc -fsanitize=address option, which is a debug tool used
to find memory problems such as leaks or use after free.

The ccan configurator is a build tool to help configure
the libnvme project for building on different platform. It
is only used during the build process and thus was never
written with making sure that it didn't leak memory.

When enabling meson's -Db_sanitize=address the build fails
because the ccan configurator, which is now compiled with
the -fsanitize=address option, crashes due to memory leaks.

This patch fixes the memory leaks in ccac configurator.

Note that the -Db_sanitize=address is simply a convenience
option used during debugging and should not be used for normal
builds. Another meson option that is available for debugging
is -Dbuildtype=debug, which disables optimization.

Signed-off-by: Martin Belanger <martin.belanger@dell.com>