Clean ccan meson infrastructure
The ccan directory is just used as another source directory
that gets built along with the other libnvme sources. There
is no point in having a meson.build in the ccan directory.
In fact, that just makes it harder to see what is actually
getting build. This commit removes the ccan meson.build file
and changes the definition of the "sources" in src/meson.build
to include the ccan files.

Signed-off-by: Martin Belanger <martin.belanger@dell.com>