Add support for Explicit Registration per TP8010.
TP8010 defines a new PDU, the Discovery Information Management
command (DIM). This PDU carries additional information that the
standard Fabrics Connect command does not provide. Among other
things, the DIM PDU allows to register the Entity Name, which
corresponds to Linux's hostname, the Entity Version (os version),
and the Host Symbolic Name (optional).

The DIM command is only sent when it is determined that the remote
discovery controller supports it. And this is indicated by a new
field in the Identify command response called the Discovery
Controller Type (dctype). This previously reserved field will be
set to 0 on pre-TP8010 discovery controllers as well as any other
types of controllers (i.e., Admin and I/O). Discovery Controllers
supporting TP8010 will set this field to 1 for Direct Discovery
Controllers and 2 for Central Discovery Controllers.

The dctype will be exposed in the sysfs by the kernel. A kernel
patch has been reviewed by several contributors and is currently
pending a merge into the kernel.

Signed-off-by: Martin Belanger <martin.belanger@dell.com>