ioctl: Fix numd calculation for corner case
numd is a 0's based dword value. Thus after dividing the lenght (in
bytes) through 4 we subtract 1. This works for all values in the range
of [3..].

By adding 4 before the division we make sure we always floor on the
next integer. With this, numd get calculated correctly for all len
values in the range [0..].

Signed-off-by: Daniel Wagner <dwagner@suse.de>