fabrics: Add nvmf_connect_disc_entry2
The current nvmf_connect_disc_entry function does not set the DHCHAP
ctrl key. The problem is that we can't really pass a new argument into
the function without breaking the existing API.

The first idea (and probably the technical correct one) was to extend
struct nvme_fabrics_config. Unfortunately, we forgot to add the
args_size member to this data structure which makes it really hard to
extend. Introducing a wrapper struct and adding new function for all
users of struct nvme_fabrics_config (nvmf_default_config,
nvmf_update_config, nvmf_add_ctrl, ...) is a bit excessive.

The main problem we want to solve here is the ability to call setters
on a newly created controller before we do a connect call. The only
place where we have this sequence is nvmf_connect_disc_entry.

Hence, this approach introduces a new nvmf_connect_disc_entry2
function which sets the key between the ctrl creation and the connect
call. We use the struct args approach so that we can extend if
needed.

Signed-off-by: Daniel Wagner <dwagner@suse.de>