python: Reformat SWIG input file
The SWIG file contains a lot of C code that was not following the
standard formatting (e.g. indent with tabs). I ran a code
beautifier to bring this code in compliance and make it easier to
read.

P.S. I verified that the code generated has not changed and tested
that eveyrhing still works as usual.

Signed-off-by: Martin Belanger <martin.belanger@dell.com>