build: Update OpenSSL wrap
Update to OpenSSL 3.0.7.

When the fallback was using v1 we never went into the v3 testing. With
updating to v3 we are going to figure out if we have LibreSSL spin of
the API. But cc.has_header_symbol() and cc.has_header() only work on
external library and not on fallbacks.

Thus if we use the fallback (internal) we just know it is API version 3
and don't do any crazy testing.

Signed-off-by: Daniel Wagner <dwagner@suse.de>