build: Move CI build steps into a scripts
Move the build instruction into a script. This allows to run these steps
also locally.

Also disable the fallback static library build as it is clearly not
working because in the dependencies rely to link against a dynamic
glibc. Instead just add a minimal static build without fallbacks.

While we are at it, also add a debug clang build.

Signed-off-by: Daniel Wagner <dwagner@suse.de>