tree: Improve ctrl lookup algorithm for TCP
As previously reported [1], there are cases where libnvme may
identify an existing controller as matching a candidate controller
when that is not really the case. In other cases, libnvme may fail
to identify an existing controller as matching a candidate
controller. This patch fixes the issue.

The problem only manifests itself when the parameters host-traddr
and/or host-iface are used. There is no problem when these parameters
are NULL for both the existing controller and the candidate controller.
Several users have reported issues when setting host-traddr and/or
host-iface to non-NULL values.

The success of the new matching algorithm depends on whether the
kernel exposes the src_addr through the sysfs (kernel 6.1 or later).
When the src_addr is exposed by the kernel, the new algorithm has
a 100% success rate. If the src_addr is not available, a few
improvements have been made to improve the success rate, but it's not
100%.

Note that to simplify the code and to make it more readable, the
function used for matching controllers has been broken into several
smaller functions. There is one function per transport type. It's
only for TCP transport that additional tests are required. The
remaining transport types have little to no changes.

[1] http://lists.infradead.org/pipermail/linux-nvme/2023-June/040295.html

Signed-off-by: Martin Belanger <martin.belanger@dell.com>