quirks: Add basic infrastructure to read external quirk database
Load a quirk database into the library when the library is loaded.

We don't expect huge numbers of entries, so let's go with a simple
array.

If we start to have many entries we can still revisit the data structure
and use some more appropriate. We don't export the table so it's an
implementation detail

Signed-off-by: Daniel Wagner <dwagner@suse.de>