tree: don't issue ns_id_desc_list if unsupported
Although the spec says Namespace Identification Descriptor list is
mandatory not all devices adhere to this. Even worse, the device logs
this as error in the log.

So if we know we are dealing with such a device just don't issue the
command in the first place. We already know it will not work.

Signed-off-by: Daniel Wagner <dwagner@suse.de>