ioctl: pass NSID in Get/Set Features commands that use it
Several features are configured on a per-namespace basis
by setting the NSID in the Set Features and Get Features commands.
But the corresponding nvme_{g,s}et_features_*() functions
aren't passing the NSID in the commands.
For the functions missing a NSID parameter, define new variants
with the NSID argument added and mark the old functions as deprecated.

Signed-off-by: Caleb Sander <csander@purestorage.com>