mi: Probe quirks on first command submission
Currently drive quirks are probed when opening a connection to the
MI endpoint. The probe itself is the only IO that's required in the
process of establishing the connection. We have an opportunity to
make connection-establishment non-blocking if we can move the probe
IO elsewhere.

We know the user is expecting IO to the endpoint when they issue a
command. Given this, exploit the call to nvme_mi_submit() to establish
the drive quirks.

Signed-off-by: Andrew Jeffery <andrew@codeconstruct.com.au>