tree: read tls_configured_key and tls_keyring from sysfs
Newer kernels expose the key which is used to setup the connection as
tls_configured_key. During operations the kernel is going to update the
tls_key thus the tls_key is not going to be useful from user point of
view.

Also the keyring used to store the keys is exposed via sysfs.

Signed-off-by: Daniel Wagner <dwagner@suse.de>