Revert various related big-endian PI commits
This reverts a series of related changes that mapped storage and
reference tags into the shared region over multiple CDWords in
big-endian ordering. Keith confirmed the CDWords are always in
host-endian ordering, and that the MSB-LSB representation in the spec
is for the metadata payload itself as handled by the driver.

Signed-off-by: Brandon Paupore <brandon.paupore@wdc.com>