Revert "fabrics: retry connect command on EINTR"
This reverts commit 2b304892292d380beaeb241e51ab2a0fdaf5c05d.

The simple approach to retry on EINTR will also prevent to exit the loop
on Ctrl-C. And adding a signal handler in the library is generally not a
good idea. It would need an additional API between users and the library
to handle signals. Let's forward EINTR and the user handle this instead.

Signed-off-by: Daniel Wagner <wagi@kernel.org>