nvme/{tree,fabrics}: do not crash when accessing disconnected controllers
When a controller is disconnected it will just clear the volatile attributes,
but remain part of the tree structure. So when traversing the tree one could
access a disconnected controller, but then c->name is NULL.
So fixup all places calling nvme_ctrl_get_name() to ensure that they don't
crash on an empty name.

Signed-off-by: Hannes Reinecke <hare@suse.de>