python: support 'with' statement for objects
The python 'with' statement allows for correct resource tracking,
where the allocated resource will be freed at the end of the statement.
This is helpful for 'host', 'subsystem', and 'controller' objects
which we might want to clean up (and clear out the internal tree).
The nice thing here is that we can do an automatic controller
disconnect, and don't have to worry about leaving stale connections
or objects around.

Signed-off-by: Hannes Reinecke <hare@suse.de>